Introduce NestedResource
A NestedResource works on an internal stack of PHP resource streams.
While creating a single target resource stream, the source resource
stream is run through until a replace pattern is found. When this
happens, the resource stream associated with the found pattern is
pushed to the stack and iterated over instead of the previous stream.